sys: anchor the buffer pool above .bss; boot to a shell
ccc emits function-scope statics (mw.c's mwinfo/mwbuf/curdrv, ...)
into .bss, which mxld places after the .data blist.  binit() sized
the buffer pool from blist to usrtop, so the 512-byte data blocks
landed on those statics and corrupted every indirect-block read —
exec loaded the wrong file.  Anchor the data at _ebss (a .bss marker
in textpad.s, linked last); the struct buf headers stay at blist.
nbuf drops 17->12, matching the 1.67 reference.

The kernel now boots init -> login -> a shell.  Stripped the
exec/rdhdr/stat/bankcopy/r_fork debug prints.

+++ src/micronix/sys/uio.c
@@ -22,7 +22,7 @@ extern struct buf *btop;            /* ditto */
  */
 struct buf *
 bget(blk, dev)
-    UINT8 blk, dev;
+    UINT blk, dev;
 {
     register struct buf *b, *f;
